iconv() returns an unsigned value so a direct comparison with -1 is invalid. (dm)
git-svn-id: svn://mielke.cc/main/brltty@5582 91a5dbb7-01b9-0310-9b5f-b28072856b6e

+++ Programs/unicode.c
@@ -113,7 +113,7 @@ getBaseCharacter (wchar_t character) {
       char outputBuffer[outputSize];
       char *outputAddress = outputBuffer;
 
-      if (iconv(handle, &inputAddress, &inputSize, &outputAddress, &outputSize) != -1)
+      if (iconv(handle, &inputAddress, &inputSize, &outputAddress, &outputSize) != (size_t)-1)
         if ((outputAddress - outputBuffer) == 1)
           return outputBuffer[0] & 0XFF;
     }
